sdk: extract owned registration planner
